+++ samples/qa-a11y/src/views/TabGroup.js
@@ -1,7 +1,7 @@
 import TabGroup from '@enact/agate/TabGroup';
+import {useCallback, useState} from 'react';
 
 import Section from '../components/Section';
-
 import appCss from '../App/App.module.less';
 
 const tabs = [
@@ -10,18 +10,29 @@ const tabs = [
 	{title: 'Theme', icon: 'display'}
 ];
 
-const TabGroupView = () => (
-	<>
-		<Section title="Default">
-			<TabGroup alt="Normal" tabPosition="before" tabs={tabs} />
-			<TabGroup alt="Disabled" disabled tabPosition="before" tabs={tabs} />
-		</Section>
+const TabGroupView = () => {
+	const [selectedNormal, setSelectedNormal] = useState(0);
+	const [selectedAriaLabelled, setSelectedAriaLabelled] = useState(0);
 
-		<Section className={appCss.marginTop} title="Aria-labelled">
-			<TabGroup alt="Aria-labelled" aria-label="This is a Label." tabPosition="before" tabs={tabs} />
-			<TabGroup alt="Aria-labelled and Disabled" aria-label="This is a Label." disabled tabPosition="before" tabs={tabs} />
-		</Section>
-	</>
-);
+	const handleSelected = useCallback(({selected}) => setSelectedNormal(selected), []);
+	const handleSelectedAriaLabelled = useCallback(({selected}) => setSelectedAriaLabelled(selected), []);
+
+	return (
+		<>
+			<Section title="Default">
+				<TabGroup alt="Normal" onSelect={handleSelected} selectedIndex={selectedNormal} tabPosition="before" tabs={tabs} />
+				<TabGroup alt="Disabled" disabled tabPosition="before" tabs={tabs} />
+			</Section>
+
+			<Section className={appCss.marginTop} title="Aria-labelled">
+				<TabGroup alt="Aria-labelled" aria-label="This is a Label." onSelect={handleSelectedAriaLabelled} selectedIndex={selectedAriaLabelled} tabPosition="before" tabs={tabs} />
+				<TabGroup
+					alt="Aria-labelled and Disabled" aria-label="This is a Label." disabled tabPosition="before"
+					tabs={tabs}
+				/>
+			</Section>
+		</>
+	);
+};
 
 export default TabGroupView;