+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The following is a curated list of changes in the Enact agate module, newest cha
 
 - `agate/TabGroup` prop `onSelect` to handle `onClick` event on it
 
+### Added
+
+- `agate/ImageItem` to have `imageItem`, `caption` and `image` publicClassnames
+- `agate/Popup` to have a `content` publicClassname
+
 ### Fixed
 
 - `agate/Panels/TabbedPanels` to not show console error when there is no children