+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The following is a curated list of changes in the Enact agate module, newest cha
 
 ### Fixed
 
+- `agate/Keypad` to fire `onChange` event with a correct value payload
 - `agate/Panels/TabbedPanels` to not show console error when there is no children
 - `agate/TabGroup` to pass `onSelect` to `ui/Group`